Verificação de dados e detalhes chatos nas solicitações
->Verificação de dados de todos os controllers parcialmente concluída.
Falta verificar 2 Requests, um em UserC e outro na HomeC.
->Bug na solicitação de amizade e de grupo:
[
Se o usuário A solicitar entrada num grupo, logo em seguida o usuário
cancela a solicitação, mas o usuário B aceita a solicitação já cancelada
(o user B [moderador] entrou na pagina quando ainda tinha a solicitação
do user A)

userA: cancelar solicitacao e solicitar de novo
userB: aceitar e recusar solicitacao

A mesma coisa pra solicitação de amizade
]
->Duplicação de solicitação de amizade
->Recusar solicitação de amizade

Dentre outras coisas...

+++ site/app/Http/Controllers/AmizadeController.php
@@ -9,55 +9,122 @@ use Illuminate\Support\Facades\DB;
 
 class AmizadeController extends Controller {
 
+   private function esseUsuarioExiste($id_user){
+      $user = \site\User::find($id_user);
+      if ($user) {
+         return true;
+      } else {
+         return false;
+      }
+   }
+
+   private function euSoliciteiEle($id_user) {
+      //Quem guarda a solicitação é quem foi solicitado, não quem enviou
+      return \site\User::find($id_user)->getSolicitacoes->contains(Auth::user()->id);
+   }
+
+   private function eleMeSolicitou($id_user) {
+      //Quem guarda a solicitação é quem foi solicitado, não quem enviou
+      return Auth::user()->getSolicitacoes->contains(\site\User::find($id_user)->id);
+   }
+
+   private function jaSaoAmigos($id_user) {
+      return Auth::user()->getAmigos->contains($id_user);
+   }
+
    public function listarAmigos() {
+      //Não precisa fazer verificação
       return view('amizades/listarAmigos', ['listaAmigos' => Auth::user()->getAmigos, 'id' => Auth::user()->id]);
    }
 
    public function listarTodosUsers() {
+      //Não precisafazer verificação
       $users = \site\User::paginate(5);
       return view('amizades/listarTodosUsers', ['users' => $users]);
    }
 
    public function listarAmigosDeOutro($id) {
-      // valida se o oto existe
+
+      //Verifica se o usuário existe
+      if($this->esseUsuarioExiste($id) == false) {
+         return "Esse usuário não existe";
+      }
 
       $user = \site\User::find($id);
       return view('amizades/listarAmigos', ['listaAmigos' => $user->getAmigos, 'id' => $id]);
-
    }
 
    public function verAmigo($id) {
+
+      //Verifica se o usuário existe
+      if($this->esseUsuarioExiste($id) == false) {
+         return "Esse usuário não existe";
+      }
+
       $user = \site\User::find($id);
-      //TODO ser usuario existe
       return view('amizades/exibirUsuario', ['user' => $user]);
    }
 
-   public function soilicitarAmizade($id) {
+   public function solicitarAmizade($id) {
+      $user = Auth::user();
 
-      if(Auth::user()->id == $id){
+      //Verifica se ele está solicitando ele mesmo
+      if($user->id == $id){
          return "Não dá pra tu ser amigo de tu mermo! kk";
       }
 
-      $eu_outro = DB::table('amizades')->where('id_user1', Auth::user()->id)->where('id_user2', $id)->get();
-      $outro_eu = DB::table('amizades')->where('id_user1', $id)->where('id_user2', Auth::user()->id)->get();
-
-      if (!$eu_outro->isEmpty() && !$outro_eu->isEmpty()) {
-         return 'mas já são amiguss';
+      //Verifica se o usuário existe
+      if($this->esseUsuarioExiste($id) == false) {
+         return "Esse usuário não existe";
       }
 
-      $user = Auth::user();
-      $destino = \site\User::find($id);
-
-      if (!$destino) {
-         return  'destino nem existe';
+      //Verifica se já são amigos
+      if ($this->jaSaoAmigos($id) == true) {
+         return 'Vocês já são amigos';
       }
+
+      //Verifica se ele já foi solicitado
+      if($this->euSoliciteiEle($id) == true) {
+         return "Essa solicitação já existe!";
+      }
+
       $solicitacoes =$user->getPedidosAmizadeEnviados();
       $solicitacoes->attach($id);
       return redirect()->back();
    }
 
+   public function cancelarSolicitacao($id_solicitado) {
+
+      //Verifica se o usuário existe
+      if($this->esseUsuarioExiste($id_solicitado) == false) {
+         return "Esse usuário não existe";
+      }
+
+      //Verifica se já são amigos
+      if ($this->jaSaoAmigos($id_solicitado) == true) {
+         return 'Vocês já são amigos';
+      }
+
+      //Verifica se a solicitação existe
+      if($this->euSoliciteiEle($id_solicitado) == false) {
+         return "Essa solicitação não existe. Talvez tenha sido recusada.";
+      }
+
+      Auth::user()->getPedidosAmizadeEnviados()->detach($id_solicitado);
+      return redirect()->back();
+   }
+
    public function aceitarAmizade($idPedinte) {
-      //valide se o pedinte existe
+
+      //Verifica se o usuário existe
+      if($this->esseUsuarioExiste($idPedinte) == false) {
+         return "Esse usuário não existe";
+      }
+
+      //Verifica se ele foi solicitado
+      if($this->eleMeSolicitou($idPedinte) == false) {
+         return "Essa solicitação não existe mais!";
+      }
 
       $pedinte = \site\User::find($idPedinte);
       $pedinte->getPedidosAmizadeEnviados()->detach(Auth::user()->id);
@@ -68,35 +135,59 @@ class AmizadeController extends Controller {
       return redirect()->back();
    }
 
+   public function recusarAmizade($idPedinte) {
+
+      //Verifica se o usuário existe
+      if($this->esseUsuarioExiste($idPedinte) == false) {
+         return "Esse usuário não existe";
+      }
+
+      //Verifica se ele foi solicitado
+      if($this->eleMeSolicitou($idPedinte) == false) {
+         return "Essa solicitação não existe mais!";
+      }
+
+      $pedinte = \site\User::find($idPedinte);
+      $pedinte->getPedidosAmizadeEnviados()->detach(Auth::user()->id);
+
+      return redirect()->back();
+   }
+
    public function listarPedidosPraMim()  {
+      //Não precisa de verificação
       return view('amizades/listarPedidosPraMim', ['lista' => Auth::user()->getSolicitacoes]);
    }
 
-   //Recebe id do amigo
    public function desfazerAmizade($id) {
       $user = Auth::user();
-      //VALIDA se o id é amg mrm
-      $user->getAmigos()->detach($id);
 
+      //Verifica se o usuário existe
+      if($this->esseUsuarioExiste($id) == false) {
+         return "Esse usuário não existe";
+      }
+
+      //Verifica se não são mais amigos
+      if ($this->jaSaoAmigos($id) == false) {
+         return 'Vocês não são mais amigos. Impossível de desfazer uma amizade que não existe mais';
+      }
+
+      $user->getAmigos()->detach($id);
       $exAmigo = \site\User::find($id);
       $exAmigo->getAmigos()->detach($user->id);
-
       return redirect()->back();
-   }
-
-   public function cancelarSolicitacao($id_solicitado) {
-      // $user = \site\User::find(Auth::user()->id);
-      // $solicitacoes = $user->getSolicitacoes()->detach($id_solicitado);
-
-         Auth::user()->getPedidosAmizadeEnviados()->detach($id_solicitado);
-
-
-      return redirect()->back();
-      // $solicitacao = \site\Solicitacao::where('id_quem_pediu', Auth::user()->id)->where('id_quem_recebeu', $id_solicitado);
    }
 
    public function perfisExternosAmigos($id) {
-      //valida se esse id ta certo
+
+      //Verifica se o usuário existe
+      if($this->esseUsuarioExiste($id) == false) {
+         return "Esse usuário não existe";
+      }
+
+      //Verifica se não são mais amigos
+      if ($this->jaSaoAmigos($id) == false) {
+         return 'Vocês não são mais amigos';
+      }
 
       $amigo = \site\User::find($id);
       $perfis = $amigo->getPerfisExternos;
@@ -104,11 +195,19 @@ class AmizadeController extends Controller {
    }
 
    public function curadoriasAmigos($id) {
-      //valida se esse id ta certo
+
+      //Verifica se o usuário existe
+      if($this->esseUsuarioExiste($id) == false) {
+         return "Esse usuário não existe";
+      }
+
+      //Verifica se não são mais amigos
+      if ($this->jaSaoAmigos($id) == false) {
+         return 'Vocês não são mais amigos';
+      }
 
       $amigo = \site\User::find($id);
       $curadorias = $amigo->getCuradorias;
-
       return view('amizades/curadoriasAmigos', ['curadorias' => $curadorias, 'amigo' => $amigo]);
    }
